requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.6
django-import-export==4.3.5
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
